+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Minimalistic fork of [mock-server/mockserver](https://github.com/mock-server/moc
 * Removed everything that is not core functionality
   * Metrics
   * MemoryTracing
-  * Templating
-  * OpenAPI 
-  * all external Body matching (XPATH, JSON, XML, ..)
+  * Templating-Engines
+  * all body matching that requires an external lib (XPATH, JSON, XML, OpenAPI, ...)
   * Version
+  * Persistence of expectations